requirements.txt
aiocache==0.12.2
aiodns==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.2.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
backcall==0.2.0
bcrypt==4.1.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
comm==0.1.3
contourpy==1.1.1
cryptography==42.0.0
cycler==0.12.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
environs==10.3.0
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.109.0
filelock==3.12.2
fonttools==4.42.1
frozenlist==1.4.0
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.3
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.5
Mako==1.3.1
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
outcome==1.2.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.6.0
pluggy==1.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.5.1
pycares==4.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.14.0
pyparsing==3.1.1
PySocks==1.7.1
pytest==7.4.0
pytest-ordering==0.6
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
pywin32==306; platform_system=="Windows"
pyzmq==25.0.2
redis==5.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqladmin==0.16.0
SQLAlchemy==2.0.25
stack-data==0.6.2
starlette==0.35.1
tornado==6.2
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.27.0
virtualenv==20.23.1
wcwidth==0.2.6
websockets==12.0
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.2
